Make HiveRecordWriter a top level class
